requirements.txt
Package          Version  
---------------- ---------
asgiref          3.2.7    
certifi          2018.8.24
click            7.1.2    
Django           3.0.6    
dnspython        1.16.0   
email-validator  1.1.1    
Flask            1.1.2    
Flask-SQLAlchemy 2.4.3    
idna             2.9      
itsdangerous     1.1.0    
Jinja2           2.11.2   
MarkupSafe       1.1.1    
pip              18.0     
pytz             2020.1   
PyYAML           5.3.1    
setuptools       40.2.0   
SQLAlchemy       1.3.17   
sqlparse         0.3.1    
Werkzeug         1.0.1    
wheel            0.31.1   
WTForms          2.3.1    
